tunefs.ocfs2: Fix resize check block_count > UINT32_MAX
The code to check resize while crossing 16TB boundaries is
already present. However, there is a bug in bit manipulation
which allowed the check to pass.

However, this needs addition checks for block64 journal options.

+++ tunefs.ocfs2/op_resize_volume.c
@@ -552,8 +552,7 @@ static errcode_t check_new_size(ocfs2_filesys *fs, uint64_t new_size,
 		}
 		try_clusters = device_clusters;
 	}
-
-	try_blocks = try_clusters << b_to_c_bits;
+	try_blocks = (uint64_t)try_clusters << b_to_c_bits;
 
 	/* Now we're guaranteed that try_clusters is within range */
